airflow/requirements.txt
apache-airflow==2.7.0
azure-storage-blob==12.19.0
pymysql==1.1.0
cryptography==41.0.7
pyarrow==14.0.1
pandas==2.1.4
python-dotenv==1.0.0
Flask==2.2.5
Werkzeug==2.3.8
pendulum<3.0
Flask-Session==0.5.0
connexion==2.14.2
gunicorn==20.1.0

